Make first line of experiment Nyles-import
I propose to make the first Python-line in every experiment-file the
import of the class Nyles instead of the class Param.  This way, if the
user forgot to activate Nyles, the error message says "failed to import
Nyles" which is a lot clearer than the previous "failed to import
Param".
This commit implements this change for the one and currently only
experiment, namely the Taylor Instability, so that this file can be
used as a tempalte for other experiments.