Advection of the buoyancy
A function to advect the buoyancy using the contravariant velocity and 
the volume. It uses the fortran_upwind function.
It was coded using the same model as bernoulli.py while I started to 
understand how everything works together.

NOT TESTED YET !
The volume is not right I think, I'm not sure that it's good to 
duplicate the buoyancy array to store the delta buoyancy out of the 
fortran routine.