Improve code without changing its functionality
With this commit, I aim to improve the existing code in the files
variables.py and vorticity.py by making it more readable, compact and
slighlty faster, with emphasis on readability.  Although this is a very
subjective matter, I hope that my changes are welcomed by the
development team.  They help me to understand what the code does and I
am sure that my changes bring the same benefits to others, too.

My changes do not modify the behaviour or functionality of the code, as
far as I am aware and as far as I have tested it.

Furthermore, I added comments and questions to the file docs/outline.txt
regarding aspects where I am not sure what the correct functionality is.