Improve the way data is stored and handled
This commit provides a considerable change in the file variables.py
resulting in a completely different way, the interaction with the data
works:

 1. The class State stores its variables as attributes (instead of,
    e.g., in a dictionary) which can be directly accessed.  There is no
    method "get" anymore, because it is not necessary.

 2. The class Vector stores its components like the elements of a
    dictionary, so that it is easily possible to loop over the
    components without using "getattr()".

Furthermore, the classes Scalar and Vector don't store the meta-
information in a dictionary "header" anymore, but directly as
attributes.  This changes also the constructor of the classes.  It makes
the whole process more transparent and gives less room for errors.

These classes now have an attribute saying whether they are prognostic
variables or not.  This will allow to write better code, e.g., in the
file timescheme.py (not yet done).

The class Scalar does not keep track of its location anymore, because
it is assumed that this is not necessary.

The attribute "size" of Scalar is now a dictionary to be consistent with
the way, how the three directions of space are usually referred to in
this project.

The model variables are now defined as NamedTuples to improve
readability.

The classes in the file variables.py have now a docstring which
describes their attributes and methods.  If this change is accepted by
the community, it can be a template for all the classes.