Make use of the "prognostic" attribute of variables
Due to the recently introduced attribute "prognostic" of Scalar and
Vector variables, it is no longer necessary to carry around a list of
prognostic variables in "param".  Instead, this list can be created
dynamically by an instance of the State class.  Furthermore, the
"duplicate" method of the class State was replaced by a method called
"duplicate_prognostic_variables", which creates a State object that
only contains prognostic variables and is therefore better adopted for
the use as a right hand side (rhs).

Two attributes of the Timescheme class which are only locally of interest
are made into local variables.  These are "timestepping" and
"timeschemelist", which are only needed in the constructor, due to a
re-structuring of the "set" method.

The class State has its "get" method back, because that seems to be
the best way to implement the class Timescheme if one wants to
preserve the versatility and modularity of each of the classes.