Computation of the Vortex Force term
This commit adds the python and fortran programs to compute the vortex 
force term :
- eps_ijk * w_j * U^k
All the data arrays are passed with the index of the vorticity upwinding 
as the inner index, to have the most efficient computation in fortran.

The fortran subroutine applies 0 velocity at the boundaries.

For now we make 6 calls to the fortran subroutine.

TODO :
- Test the subroutine to make sure that it computes what is expected
- Make it more efficient by calling the fortran subroutine 3 times 
instead of 6
  (This should be done once we know that the algorithm is correct)