Correct for missing 1/vol-term in advection model
In the upwind scheme used for the advection of a tracer, it is necessary
to divide by the volume to get dimensions (and values) correct.  This
will become important if we use grids with non-uniform cell-volumes.
The effect of this correction can already be observed by setting the
variable vol in tracer.py to a non-unity value.