Add proposal for the implementation of model_les
The advantages and disadvantages of the proposed implementation of
the model class together with the required changes of the timescheme
are discussed in the docstring of the file.

The implementation is not final.  It lacks most importantly the script
for solving the Poisson equation for pressure.  Nevertheless, a dummy
run of the model can be made.