Provide a multigrid
This commit is a merge of a local branch. Things are likely
broken. The multigrid worked in the mono core case -- tested. The
whole structure is designed to handle the multi core case.

This commit should be followed by a small one, for which all broken
pieces will have been fixed