Provide some first simple experiments
The vortex experiments provided with this commit can be used to test if
the simulation results of Nyles are consistent with 2D experiments and
if basic fluid mechanics works.  The vertical experiments can give
interesting results, too.

Furthermore, the provided experiment file can be used as a template for
other experiments.