Add option to choose between automatic or fix timesteps
A new parameter is added to param which lets the user choose between
a fixed and an automatically chosen timestep dt.  In the Nyles class,
functionality to handle these two cases is implemented.

The three experiments are adapted to this new option.

Also a new method is added to nylesIO which is called at the end of the
simulation to save the final state if necessary.