Rewrite vortex_force with 'flipview' method
The code looks clearer to me like this. This is how I realized that we
have a bug here, because, thanks to Alex comment, l59
one term had indeed omega_i with i the inner index = correct
but the other term is omega_j and j is not the inner index.