Add class for handling of user parameters
This commit re-writes the file parameter.py under the more suitable name
parameters.py, which now provides a class to handle the parameters that
the user can modify.  The default values of the parameters are, as
usual, given in defaults.json.  The user is not obligated to re-list
parameters that he does not intend to change.  This makes the
experiments sometimes shorter and always clearer.

Parameters are sorted in categories (currently 5) to make it easier for
the user to understand the parameters.  This distinction is not used
when the parameters are given as a dictionary to any other class, like
grid or IO.

New typo prevention: If the user tries to set a parameter that does not
exist, they see an error.

Parameters that are automatically set are not controlled by the new
class UserParameters.

Every (working) experiment is adapted to the new parameter handling.
Also the Nyles class is adjusted.

A new paramter dt_max is added to have consistent behaviour for
velocities close to zero.

The import "nyles" was renamed "nyles_module" in the buoyancy
instability experiment due to a double usage of the name "nyles".

The "timing" modifier is added to the compute_dt method.

TODO: the parameters for multigrid and MPI are currently added in the
Nyles class.  There might be a better place for them.