Implement usage of the metric tensor
Dummy values for dx,dy,dz and related metric quantities are replaced by
their actual values given by the Grid class.  References to a Grid
object are added were necessary.

The grid class is extended by two dictionaries which make the
implementations in other places simpler.

The usage of "modelname" in the Nyles class is adjusted to the values
given in defaults.json.