Remove unnecessary volume-term from Cartesian advection model
The volume appears only in the advection of a tracer if non-uniform
coordinates are used.  In that case, however, there are several volumes
and each of them is a 3D array.  So it makes sense to handle the simple
case of uniform volumes separate.  In this case, the volume term does
not appear in the advection equation.  It is removed with this commit.