Add missing reset of dstate + fix metric-term in Bernoulli
The arrays in dstate must be reset to zero on every iteration of the
time integration.  This is implemented in this commit in a crude way.
Result:  the linear wave stays stable!
But:  the velocity field is still not divergence free.
TODO:  check if this method takes too much time and if so, put the reset
into the existing Fortran code, as mentioned in the new comment in the
rhs function of the LES model.

This commit also fixes a missing usage of the metric-term dz in the
fortran Bernoulli script.
Result:  the experiment can be run with any domain size and resolution!