Provide a working experiment with linear dynamics
The experiment shows a wave on a density-interface oscillating stably.
It is implemented on a 64x64x64-grid with 5th order advection and
3rd order timestepping, taking only linear dynamics into consideration.

Note: the lenghts of the domain must be such that dx, dy, dz are all
equal to 1; otherwise there is a problem in the pressure calculation.

The plotting command is removed from the experiment, because there
exists by now a separate plotting module.

Noise is removed from the experiment since it is not necessary.

Further changes:

 - Make the run-method of Nyles a bit more verbose/informative.

 - Correct the documentation of two user-parameters.