New Features: Linear Dynamics + Disk Space Checks
The linear approximation so far used in the LES model has proven to give
interesting results.  Therefore this linear model should be preversed.
It has now its own modelname "linear" and an experiment demonstrating
the model (which is basically the old buoyancy instability experiment
with the correct name "linear_wave").  The linear model is seamlessly
integrated into the LES model, which now has its full functionality
back.  Possible questions that can be addressed with the new model
include:

 - what is the difference between the linearized momentum equations and
 the full momentum equations?

 - why/where/when are the non-linear terms in the equation necessary?

It is quite easily possible to imagine interesting extensions to this,
like a model linearized around a steady flow (instead of a stationary
situation).

With this change, the buoyancy experiment is made back into an actual
unstable situation using the (non-linearized) LES model.

Another new feature is the low-disk-space warning.  If the available
disk space in the output directory drops below a limit, the simulation
is paused, so that the user can decide whether he wants to continue,
free some space, or stop the experiment.  The limit is configurable in
the parameters set by the user.  This is necessary due to the huge file
sizes that Nyles can create in short time, especially if all model
variables are saved.

An unused function is removed from model_les_AL; the function has
already been implemeneted in the projection module.

The scaling in the milk-coffee experiment is corrected.

The vortices experiment is named vortex to be consistent with the other
experiments.