Provides the first stable full nonlinear LES
Yes! the main change is a complete rewrite of vortex_force. It was a
big mess between conventions and indices. A second change is to add a
function to diagnose all the diagnostic variables after the prognostic
ones have been updated.

The interfacial wave is now stable, including with 5th order and
LFAM3 (Euler and 5th order are unstable but this is expected)

I have restored the Taylor instability and ... that's beautiful!!!