Implement consistent handling of units and dimensions
Two new parameters are added which set the unit used for duration and
length in the model.  By default, they are "s" for second and "m" for
meter.  The user can change them to whatever is suitable for the
experiment, e.g. kilometer and year.  The correct units for all the
phyiscal quantities will be saved in the history file.

Thanks to this and to the improved multigrid, the coffee cup experiment
is implemented with realistic sizes.  This experiment uses CGS units
which, although unpopular, are well suited for this purpose.

An error is corrected which assigned pressure the unit m/s², but it
should be m²/s².

The Coriolis parameter is now allowed to take any real value.  Negative
values are necessary to simulate situations on the southern hemisphere,
0 is not necessary but useful to make automatization simpler.

The special treatment of the "work" array is removed since this array
does not exist anymore.