Add passive tracers and richer animations
A new parameter allows to add an arbitrary number of passive tracers.
They are advected like buoyancy without explicit diffusion but do not
interact with the model in any other way.  Thus they are passive.  They
have the simple names t0, t1, ... where "t" stands for "tracer".

This allows to visualise flow patterns and mixing processes in a much
richer way.  An example is provided as the video for the coffee cup
experiment.

Tracers are automatically saved in the history file, because they are
only of interest in the analysis of the data after the simulation
run (and possibly in the live plotting).

Further options were added to customise the live plotting.  These
options are shown in the linear wave and coffee cup experiments.

Note however, that the 3D animation still slows down the simulation.
Thus it is usually better to disable live plotting and visualise the
history file later (or convert it into a video) with the animate tool.