Put backup of script file at the right place
The backup of the script file must be done after the output directory
has been created.  Otherwise cp fails.

Therefore, I think it makes sense to move the backup method to the
NylesIO class, which creates the output directory.  NylesIO also
prevents overwriting the output file when the parameter "count" is used
instead of "overwrite".  So the implementation is shortened.

I change the usage of subprocess to shutil to make it platform
independent.