Prevent the accidental change of parameters after initialisation
Implement a small, new routine that warns the user if they try to modify
a parameter after the Nyles object is created.

Move back the call to "backup the script" to the Nyles class (while the
logic of the backup stays in the NylesIO class).  I think this will be
better to generalise for program runs on multiple processors.

Improve the "check free disk space" routine:  If the program stops due
to insufficient disk space and the user then frees up disk space before
continuing the program, the checks stay on and are not turned off.

Add a parameter to update the animation not after every iteration to
speed up the routine.