Update "bugs" folder
A new bug report is added for the problem discussed today, that the
boundary conditions must be handled with more care.

The bug report for the advection issue is deleted; this was not a bug.
The helpful comment by Alex is added for future reference to the
corresponding Issue on GitHub and the issue is closed there, too.