Give the two LES models proper names
The file with the simple name model_les is the one to be used.

The alternative approach was given the suffix "_alt" (for alternative).
Importing it raises a NotImplementedError to prevent imprudently
replacing the real LES model by the non-functional version.  This file
is kept, however, since it might be finished in the future or may be
used as a starting point to develop other models.