Improve NylesIO
The new parameter simplified_grid provides an option so save all
quantities at cell centers in the history file.  This makes it easier to
look at the file with ncview.  Of course it does not change the
calculations.  The coffee cup experiment makes use of this new feature.

One can argue whether this value should be True or False by default.  It
is currently False to stay consistent and to provide realistic results.
However, it might be desirable to activate this new feature by default
since the output file looks so much more readable.

The parameter variables_in_history now takes also the values
"prognostic" to save all prognostic variables and "p+p" to save all
prognostic variables and pressure.  The reason for these options is that
the diagnostic variables are quite inexpensive to calculate and thus can
be calculated as needed instead of blowing up the history file.
Pressure is a special case, since it is diagnostic but not very easy to
calculate from the history file.

The structure of NylesIO was improved and simplified without changing
its interfaces (i.e. public methods and attributes).