Finish up parameter handling for multigrid
Remove the lines in nyles.py which set the multigrid parameters because
they are now set by the user.

Remove "topology" from param, because it is never used.  If it is
needed, one can use "geometry".  Otherwise one can use topo.topology, as
it is done in most places right now.

Add "myrank" to param.  This is useful in one place but might be also
useful for the modelgrid and the IO class.

Change the availability string of the mg params to match their datatype.