Introduce explicit diffusion on the tracer dynamics
The diffusion coefficients are defined by the user as a dictionary
See experiment/jetalongx.py

param.physics['diff_coef'] = {'b': 1e-1}

of course, we add as many coefficients as tracers.

This development anticipates the need to have explicit diffusivities
and viscosities to have control the Prandtl number and the Peclet
number.

TODO: add the diffusion on momentum (the covariant component). Good
news: the operator is exactly the same as for tracers. The
fortran_dissipation can be called as is.

The diffusion computation is done in a Fortran routine but clearly
we want to blend that computation in fortran_upwind (TODO). It will
slightly less readable but it spares three 3D loops. For momentum, we
can't do that! we need to call the fortran_dissipation.