Attempt to fix the BC issue
add one ghostpoint to Scalar arrays, on their left, if there is a wall
on the left (see bugs/boundary_conditions.txt for explanations)

add a boundarycond.py routine to handle BC on velocity and vorticity

If this development is correct (after having been carefully tested)
then we'll optimize the coding. As it is now, there are too many
.view() in boundarycond.py. Besides, it is smarter to impose

       u = state.u['i'].flipview('i')
       u[0, :, :] = 0.

rather than

       u = state.u['i'].view('i')
       u[:, :, 0] = 0.

do you see why?