Continued the work on free-slip BC, in the closed case I chose to use the cells around the domain as ghostpoints Doesn't work yet, there are weird oscillations around the boundaries. Those oscillations pull buoyancy from the ghostpoints
- Removed the additional cell that was added in topology.py because it 
was breaking the computation (I don't know why)
So I chose to use the cells around the domain as ghostpoints. I didn't 
try this with MPI yet, but it should be ok.
Maybe there is a better way to do this ?
- Added BC on all velocity directions, so there is no need to add 
boundaries on vorticity anymore
- In tracer.py : set i0 to 0 always. Didn't change fortran_upwind so we 
can easily come back to the previous state

- Tried to use pressure continuity at the boundaries : didn't work but I 
still left it as a comment in projection.py

Happy new year !