Fix an important multigrid issue
mg was no longer converging in the case with subdomains. The culprit
was that the interpolation matrix should also interpolate in the halo
so that Acoarse=R*Afine*I is correctly defined on all inner
points. The lack of convergence was making U divergent along inner
subdomains boundaries, causing spurious creation/destruction of buoyancy

join is a little bit more flexible (with varname)

ran a pep8 on grids.py