Restore back the non-conservative form of the vortex-force term
In anticipation of enforcing the BC on momentum via the the tangential
components of the vorticity, the interpolation of vorticity now uses
the value at the right boundary (it should also pick the value at the
left boundary but the vorticity arrays are currently too short by one
grid point!).

The vortex-force term interpolates the vorticity, and not the
vorticity*U. In the latter case, since W=0 at the top boundary, the
surface vorticity can't be entrained downward. This aspect of the
model will definitely be one of the originalty.